+++ CHANGELOG.md
@@ -9,10 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.2.0](https://github.com/fivetran/terraform-provider-fivetran/compare/v1.1.27...v1.2.0)
 
+## Added
 - New resource `fivetran_local_processing_agent` that allows to manage Local Processing Agents.
 - New data source `fivetran_local_processing_agent` that allows to retrieve details of the existing Local Processing Agent for a given identifier.
 - New data source `fivetran_local_processing_agents` that allows to retrieve the list of existing Local Processing Agents available for the current account.
 
+## Updated
 Updates to support management of local processing agents:
 - Resource `fivetran_connector` updates:
   - Added field `fivetran_connector.local_processing_agent_id`.
@@ -28,10 +30,16 @@ Updates to support management of local processing agents:
   - Added field `fivetran_connector.local_processing_agent_id`.
   - Added field `fivetran_connector.networking_method`.
 
-## [1.1.27](https://github.com/fivetran/terraform-provider-fivetran/compare/v1.1.26...v1.1.27)
+- Resource `fivetran_connector_schema_config` reworked, added new validation logic and new field `validation_level`.
+    - `fivetran_connector_schema_config.validation_level` allows to setup desired level of schema validation before apply
+        - NONE: no validation needed, the fastest way to apply schema, especially for a newly created connector - it will use new [Create a Connector Schema Config](https://fivetran.com/docs/rest-api/connectors#createaconnectorschemaconfig) endpoint.
+        - TABLES: validate only schemas and tables names. This level will require only schema reloading, without requests to retrieve columns for every configured table
+        - COLUMNS: performs full configuration validation, but requires to make requests to the source to retrieve all columns for every table.
+- Resource `fivetran_connector_schema_config` won't fail if some parts of schema were deleted from source and now able to handle such changes. 
+    - Provider will show warnings in output if some configures schema element doesn't exist anymore in source schema.
+    - Provider will fail on attempt to patch not existing schema, table or columns if appropriate `validation_level` specified and will prompt to update resource configuration.
 
 ## Fixed
-
 - Remove Default value from `networking_method` field in `fivetran_connector`
 
 ## [1.1.26](https://github.com/fivetran/terraform-provider-fivetran/compare/v1.1.25...v1.1.26)